Refactor code, move posts list to seperate component #14
